requirements.txt
argparse
git+https://github.com/m-bain/whisperx.git
git+https://github.com/openai/whisper.git
python-dotenv
